Add files via upload
This is your First Milestone Project. We would be exploring Data of Titanic survivors. Although the dataset is pretty common on kaggle but the activities to be performed in this Project Assignment are unique to our curriculum and entirely based on the topics and activities performed in class. There are two notebooks attached here which are based on the following activities performed during the class:
- Uni-variate Analysis
- Multivariate Analysis
- Data Pre-processing
- Data Visualization
Please attempt univariate_anaysis first before moving on to multivariate_analysis.